ovn_northd.dl: Fix DDlog tag allocation logic
It appears that the C version supports two methods of static VLAN tag
assigninment: by setting the `tag_request` field, which gets copied
to `tag` by northd, and by assigning the `tag` field directly.  The
former has precedence over the latter.  This seems a bit hairy, but
the test suite expects this behavior, so we now support it in DDlog.