ovn-northd-ddlog: Untested implementation of `ddlog_handle_reconnect`
ovn-northd-ddlog does not currently handle OVSDB reconnects (when DB
connection is dropped and later restored) correctly.  It should detect
such events, clear all input tables in the affected database and
re-populate them with a complete database snapshot received from
OVSDB.

This commit adds an untested implementation of the clear-and-repopulate
logic in the `ddlog_handle_reconnect`.